extract exchange prediction into a standalone function and move it to a separate file
