+++ core/callinfo/collector_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ftl/conval"
 	"github.com/ftl/hamradio/callsign"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -96,90 +95,6 @@ func TestCollector_addDXCC(t *testing.T) {
 			} else {
 				assert.Equal(t, test.expectedDXCCPrefix, info.DXCCEntity.Prefix, "DXCC prefix")
 			}
-		})
-	}
-}
-
-func TestCollector_predictExchange(t *testing.T) {
-	rstProperty := core.ExchangeField{
-		Field:            "theirExchange_report",
-		Properties:       conval.ExchangeField{conval.RSTProperty},
-		CanContainReport: true,
-	}
-	nameProperty := core.ExchangeField{
-		Field:            "theirExchange_name",
-		Properties:       conval.ExchangeField{conval.NameProperty},
-		CanContainReport: true,
-	}
-
-	tests := []struct {
-		name                string
-		theirExchangeFields []core.ExchangeField
-		dxcc                dxcc.Prefix
-		qsos                []core.QSO
-		currentExchange     []string
-		historicExchange    []string
-		expected            []string
-	}{
-		{
-			name:                "only report, the entry field must be initialized with the default report",
-			theirExchangeFields: []core.ExchangeField{rstProperty},
-			currentExchange:     []string{"59"},
-			historicExchange:    []string{""},
-			expected:            []string{"59"},
-		},
-		{
-			name:                "name, empty",
-			theirExchangeFields: []core.ExchangeField{nameProperty},
-			currentExchange:     []string{""},
-			historicExchange:    []string{""},
-			expected:            []string{""},
-		},
-		{
-			name:                "name, only current",
-			theirExchangeFields: []core.ExchangeField{nameProperty},
-			currentExchange:     []string{"Flo"},
-			historicExchange:    []string{""},
-			expected:            []string{"Flo"},
-		},
-		{
-			name:                "name, with history",
-			theirExchangeFields: []core.ExchangeField{nameProperty},
-			currentExchange:     []string{""},
-			historicExchange:    []string{"Flo"},
-			expected:            []string{"Flo"},
-		},
-		{
-			name:                "name, history over current",
-			theirExchangeFields: []core.ExchangeField{nameProperty},
-			currentExchange:     []string{"Hans"},
-			historicExchange:    []string{"Flo"},
-			expected:            []string{"Flo"},
-		},
-		{
-			name:                "name, qso over history",
-			theirExchangeFields: []core.ExchangeField{nameProperty},
-			qsos:                []core.QSO{{TheirExchange: []string{"Steve"}}},
-			currentExchange:     []string{""},
-			historicExchange:    []string{"Flo"},
-			expected:            []string{"Steve"},
-		},
-		{
-			name:                "name, history over unclear qso",
-			theirExchangeFields: []core.ExchangeField{nameProperty},
-			qsos:                []core.QSO{{TheirExchange: []string{"Steve"}}, {TheirExchange: []string{"Bud"}}},
-			currentExchange:     []string{""},
-			historicExchange:    []string{"Flo"},
-			expected:            []string{"Flo"},
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			c := &Collector{
-				theirExchangeFields: test.theirExchangeFields,
-			}
-			actual := c.predictExchange(test.dxcc, test.qsos, test.currentExchange, test.historicExchange)
-			assert.Equal(t, test.expected, actual)
 		})
 	}
 }
